+++ satella/instrumentation/metrics/structures/__init__.py
@@ -1,2 +1,4 @@
 from .threadpool import MetrifiedThreadPoolExecutor
 from .cache_dict import MetrifiedCacheDict
+
+__all__ = ['MetrifiedCacheDict', 'MetrifiedThreadPoolExecutor']