+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * added `default_value_factory` to `CacheDict`
 * added `unique`
 * `MetrifiedCacheDict`
+* added `CacheDict.has_info_about`